requirements.txt
ansicon==1.89.0; python_full_version >= "3.6.2" and python_version < "4" and python_version >= "2.7" and platform_system == "Windows"
arrow==1.2.3; python_full_version >= "3.6.2" and python_version < "4" and python_version >= "3.6"
asgiref==3.5.2; python_full_version >= "3.6.2" and python_version < "4" and python_version >= "3.8"
blessed==1.19.1; python_full_version >= "3.6.2" and python_version < "4" and python_version >= "2.7"
brotli==1.0.9; python_version >= "3.7"
certifi==2022.9.24; python_version >= "3.7" and python_version < "4"
charset-normalizer==2.1.1; python_version >= "3.7" and python_version < "4" and python_full_version >= "3.6.0"
croniter==1.3.7; (python_version >= "2.6" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
dj-database-url==1.0.0
django-appconf==1.0.5; python_version >= "3.6"
django-compressor==4.1
django-debug-toolbar==3.7.0; python_version >= "3.7"
django-picklefield==3.1; python_full_version >= "3.6.2" and python_version < "4" and python_version >= "3"
django-q==1.3.9; python_full_version >= "3.6.2" and python_version < "4"
django==4.1.3; python_version >= "3.8"
factory-boy==3.2.1; python_version >= "3.6"
faker==15.3.1; python_version >= "3.7"
feedparser==6.0.10; python_version >= "3.6"
freezegun==1.2.2; python_version >= "3.6"
gunicorn==20.1.0; python_version >= "3.5"
idna==3.4; python_version >= "3.7" and python_version < "4"
jinxed==1.2.0; python_full_version >= "3.6.2" and python_version < "4" and python_version >= "2.7" and platform_system == "Windows"
psycopg2==2.9.5; python_version >= "3.6"
pyjwt==2.6.0; python_version >= "3.7"
python-dateutil==2.8.2; python_full_version >= "3.6.2" and python_version >= "3.6" and python_version < "4" and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.7")
python-dotenv==0.20.0; python_version >= "3.5"
rcssmin==1.1.0
redis==3.5.3; python_full_version >= "3.6.2" and python_version < "4"
requests==2.28.1; python_version >= "3.7" and python_version < "4"
rjsmin==1.2.0
rollbar==0.16.3
sgmllib3k==1.0.0; python_version >= "3.6"
six==1.16.0; python_full_version >= "3.6.2" and python_version < "4" and python_version >= "3.6"
sqlparse==0.4.3; python_full_version >= "3.6.2" and python_version < "4" and python_version >= "3.8"
tzdata==2022.6; sys_platform == "win32" and python_version >= "3.8" and python_full_version >= "3.6.2" and python_version < "4"
urllib3==1.26.12; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.7"
wcwidth==0.2.5; python_full_version >= "3.6.2" and python_version < "4" and python_version >= "2.7"
whitenoise==6.2.0; python_version >= "3.7"
